// src/es/schemas/ingest.ts
export const IngestDocument = z.object({
  _id: z.lazy(() => Id).describe('Unique identifier for the document. This ID must be unique within the `_index`.').optional(),
  _index: z.lazy(() => IndexName).describe('Name of the index containing the document.').optional(),
  _source: z.any().describe('JSON body for the document.')
}).meta({ id: 'IngestDocument' })
export type IngestDocument = z.infer<typeof IngestDocument>

export const IngestPipeline = z.object({
  description: z.string().describe('Description of the ingest pipeline.').optional(),
  on_failure: z.array(z.lazy(() => IngestProcessorContainer)).describe('Processors to run immediately after a processor failure.').optional(),
  processors: z.array(z.lazy(() => IngestProcessorContainer)).describe('Processors used to perform transformations on documents before indexing. Processors run sequentially in the order specified.').optional(),
  version: z.lazy(() => VersionNumber).describe('Version number used by external systems to track ingest pipelines.').optional(),
  deprecated: z.boolean().describe('Marks this ingest pipeline as deprecated. When a deprecated ingest pipeline is referenced as the default or final pipeline when creating or updating a non-deprecated index template, Elasticsearch will emit a deprecation warning.').optional(),
  _meta: z.lazy(() => Metadata).describe('Arbitrary metadata about the ingest pipeline. This map is not automatically generated by Elasticsearch.').optional(),
  created_date: z.lazy(() => DateTime).describe('Date and time when the pipeline was created. Only returned if the `human` query parameter is `true`.').optional(),
  created_date_millis: z.lazy(() => EpochTime).describe('Date and time when the pipeline was created, in milliseconds since the epoch.').optional(),
  modified_date: z.lazy(() => DateTime).describe('Date and time when the pipeline was last modified. Only returned if the `human` query parameter is `true`.').optional(),
  modified_date_millis: z.lazy(() => EpochTime).describe('Date and time when the pipeline was last modified, in milliseconds since the epoch.').optional(),
  field_access_pattern: z.lazy(() => IngestFieldAccessPattern).describe('Controls how processors in this pipeline should read and write data on a document\'s source.').optional()
}).meta({ id: 'IngestPipeline' })
export type IngestPipeline = z.infer<typeof IngestPipeline>